Create LWC bundle to hold JS utility functions that can be referenced from other LWCs
